[PE-6452] Refactor web lineup tiles
