--- packages/web/src/components/track/desktop/ConnectedTrackTile.module.css
@@ -1,101 +1,0 @@
-.container {
-  --border-width: 1px;
-  position: relative;
-  background: var(--harmony-white);
-  border: var(--border-width) solid var(--harmony-n-100);
-  border-radius: 8px 8px 0px 0px;
-  display: inline-flex;
-  width: 100%;
-  cursor: pointer;
-}
-
-.container.standalone {
-  transform: scale3d(1, 1, 1);
-  border-radius: 8px;
-  box-shadow:
-    0px 2px 5px var(--tile-shadow-3),
-    0px 1px 0px var(--tile-shadow-2),
-    0px 0px 1px var(--tile-shadow-1);
-}
-
-.container.standalone:not(.loading):not(.uploading):hover {
-  background: var(--harmony-n-25);
-  box-shadow:
-    0 1px 5px 1px var(--tile-shadow-1-alt),
-    0 1px 0 0 var(--tile-shadow-2),
-    0 2px 10px -2px var(--tile-shadow-3);
-  transform: scale3d(1.005, 1.005, 1.005);
-}
-
-.container.standalone:not(.loading):not(.uploading):active {
-  box-shadow:
-    0 0 1px 0 var(--tile-shadow-1),
-    0 2px 3px -2px var(--tile-shadow-3);
-}
-
-.container.large.standalone {
-  margin-bottom: 12px;
-}
-
-.container.small.standalone {
-  margin-bottom: 8px;
-}
-
-.container.large {
-  height: 154px;
-}
-
-.container.small {
-  height: 124px;
-}
-
-.container:hover .artworkIcon {
-  opacity: 0.75;
-}
-
-.active .name {
-  color: var(--harmony-primary);
-}
-
-.iconKebabHorizontal path {
-  fill: var(--harmony-n-400);
-}
-
-.menuContainer {
-  height: 24px;
-  width: 19px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  transform: scale3d(1, 1, 1);
-  transition: all 0.07s ease-in-out !important;
-}
-
-.menuContainer:hover {
-  transform: scale3d(1.1, 1.1, 1.1);
-}
-
-.menuContainer:active {
-  transform: scale3d(0.95, 0.95, 0.95);
-}
-
-.menuKebabContainer {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  align-items: center;
-}
-
-.menuKebabContainer.large {
-  height: 42px;
-  min-width: 38px;
-}
-
-.menuKebabContainer.small {
-  height: 38px;
-  min-width: 38px;
-}
-
-.badgeWrapper {
-  margin-left: 4px;
-}
