--- packages/web/src/components/track/desktop/PlaylistTile.module.css
@@ -1,34 +1,0 @@
-.large,
-.medium {
-  margin-bottom: 16px;
-}
-
-.small {
-  margin-bottom: 8px;
-}
-
-.moreTracks {
-  background: var(--harmony-bg-surface-1);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  color: var(--harmony-n-400);
-  font-size: 12px;
-  font-weight: 500;
-  height: 27px;
-  cursor: pointer;
-}
-
-.moreTracks:hover {
-  background: var(--harmony-n-50);
-}
-
-.moreArrow {
-  margin-left: 4px;
-  height: 16x;
-  width: 16px;
-}
-
-.moreArrow path {
-  fill: var(--harmony-n-400);
-}
