+++ packages/web/src/components/track/desktop/TrackTile.module.css
@@ -1,148 +1,3 @@
-.container {
-  --border-width: 1px;
-  position: relative;
-  background: var(--harmony-white);
-  border: var(--border-width) solid var(--harmony-n-100);
-  border-bottom: 0;
-  border-radius: var(--harmony-unit-2) var(--harmony-unit-2) 0px 0px;
-  display: inline-flex;
-  width: 100%;
-}
-
-.container.standalone {
-  border-radius: var(--harmony-unit-2);
-}
-
-.container.large.standalone {
-  margin-bottom: var(--harmony-unit-4);
-}
-
-.container.small.standalone {
-  margin-bottom: var(--harmony-unit-3);
-}
-
-.container.large {
-  height: 144px;
-}
-
-.container.small {
-  height: 128px;
-}
-
-.order {
-  flex: 0 0 var(--harmony-unit-8);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  font-weight: var(--harmony-font-bold);
-  font-size: var(--harmony-font-m);
-  color: var(--harmony-n-600);
-  position: relative;
-}
-
-.crownContainer {
-  position: absolute;
-  top: var(--harmony-unit-12);
-}
-
-.crownContainer svg {
-  width: 18px;
-  height: 13px;
-}
-
-.crownContainer path {
-  fill: var(--harmony-n-600);
-}
-
-.large .imageContainer {
-  flex: 0 0 128px;
-  height: 128px;
-  margin: var(--harmony-unit-2) 0px;
-}
-
-/* If the ordering is not present, add extra space to the left */
-.large .imageContainer.leftSpacing {
-  margin-left: var(--harmony-unit-2);
-}
-
-.small .imageContainer {
-  flex: 0 0 108px;
-  height: 108px;
-  margin: var(--harmony-unit-2) 0px;
-}
-
-/* If the ordering is not present, add extra space to the left */
-.small .imageContainer.leftSpacing {
-  margin-left: var(--harmony-unit-2);
-}
-
-.container:hover .artworkIcon {
-  opacity: 0.75;
-}
-
-.container.isDisabled {
-  cursor: default;
-}
-
-.body {
-  display: inline-flex;
-  flex-direction: column;
-  overflow: hidden;
-  flex: 1 1 auto;
-  padding: var(--harmony-unit-2) var(--harmony-unit-3) var(--harmony-unit-2)
-    18px;
-}
-
-.text {
-  overflow: hidden;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-}
-
-.volumeIcon {
-  margin-left: 6px;
-}
-
-.isHidden {
-  visibility: hidden;
-}
-
-.socialsRow {
-  display: flex;
-  gap: var(--harmony-unit-4);
-}
-
-.large .socialsRow .skeleton {
-  height: var(--harmony-unit-6);
-}
-
-.small .socialsRow .skeleton {
-  height: 18px;
-}
-
-.large .socialsRow {
-  height: var(--harmony-unit-6);
-  margin-right: 80px;
-}
-
-.small .socialsRow {
-  height: 18px;
-  margin-right: var(--harmony-unit-20);
-}
-
-.small .bottomRight {
-  /* Match socialsRow to keep these in line */
-  height: 18px;
-}
-
-.socialInfo {
-  display: inline-flex;
-  flex: 1 1 100%;
-  min-width: 0;
-  justify-content: flex-start;
-}
-
 /* Right side of the track tile */
 .topRight {
   position: absolute;
@@ -158,24 +13,6 @@
   letter-spacing: 0.2px;
 
   /* Neutral Light 4 [day] */
-  color: var(--harmony-n-400);
-}
-
-.topRightIcon {
-  height: var(--harmony-unit-4);
-  width: var(--harmony-unit-4);
-  margin-right: 6px;
-}
-
-.topRightIcon path {
-  fill: var(--harmony-n-400);
-}
-
-.topRightIconLabel {
-  display: flex;
-  align-items: center;
-  flex: 1 0 auto;
-
   color: var(--harmony-n-400);
 }
 
@@ -213,82 +50,4 @@
 
 .completeIcon path {
   fill: var(--harmony-secondary);
-}
-
-.bottomRight {
-  bottom: 0;
-  position: absolute;
-  right: 0;
-  display: inline-flex;
-}
-
-.plays {
-  text-align: right;
-  letter-spacing: 0.2px;
-  color: var(--harmony-n-400);
-}
-
-/* Track Tile Divider */
-.divider {
-  height: 1px;
-  min-height: 1px;
-  width: 100%;
-  background: var(--harmony-n-100);
-}
-
-.large .divider {
-  margin: var(--harmony-unit-2) 10px var(--harmony-unit-2) 0px;
-}
-
-/* If Header (Playlist or Album) is not present, add more spacing up top */
-.large .withoutHeader .divider {
-  margin: var(--harmony-unit-2) 10px var(--harmony-unit-2) 0px;
-}
-
-.small .divider {
-  margin: var(--harmony-unit-1) 0px 6px 0px;
-}
-
-/* Bottom Row Track TIle */
-.iconButton {
-  display: flex;
-  height: var(--harmony-unit-6);
-  width: var(--harmony-unit-6);
-  align-items: center;
-  transform: scale3d(1, 1, 1);
-  transition: all 0.12s ease-in-out !important;
-}
-
-.iconButton:hover {
-  transform: scale3d(1.18, 1.18, 1.18);
-}
-
-.iconButtonContainer.isDisabled .iconButton:hover {
-  transform: scale3d(1, 1, 1);
-}
-
-.iconButton:active {
-  transform: scale3d(0.95, 0.95, 0.95);
-}
-
-.iconButtonContainer.isDisabled .iconButton:active {
-  transform: scale3d(1, 1, 1);
-}
-
-.iconShareContainer {
-  height: var(--harmony-unit-6);
-  width: var(--harmony-unit-6);
-}
-
-/* Make the icon buttons a larger click target */
-.iconButton:after {
-  content: '';
-  padding: var(--harmony-unit-4);
-  position: absolute;
-  left: var(--negative-unit);
-  top: var(--negative-unit);
-}
-
-.iconKebabHorizontal path {
-  fill: var(--harmony-n-400);
 }
